requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
datasets==2.7.1
dill==0.3.6
filelock==3.8.0
frozenlist==1.3.3
fsspec==2022.11.0
huggingface-hub==0.11.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.14
nltk==3.7
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pyarrow==10.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rouge-metric==1.0.1
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch-cluster==1.6.0+pt113cpu
torch-geometric==2.1.0.post1
torch-scatter==2.1.0+pt113cpu
torch-sparse==0.6.15+pt113cpu
torch-spline-conv==1.2.1+pt113cpu
torchaudio==0.13.0+cpu
torchvision==0.14.0+cpu
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.13
xxhash==3.1.0
yarl==1.8.1
